Begun authoring the handling of an "Untyped Plutus program by building the structs that are to store the hierarchy of the program's fucntional programming execution.
The struct "inupTerm_s" is used as the generic term as defined in the Plutus Core
specification document's Fig. 25. This struct is made from a payload union of
structs and an enumeration variable indicating what term this is. An initializer
function was added.

The struct "inupConstant_s" is presently abscent (Fig. 24).
IN 2021/11/27