Corrected the struct members in "inup.h" to use the correct pointer to a "Term" object (instead of the nonexistent) structs that I had come up with).
Corrected the lambda abstraction's "name" to be an integer that represents
a variable in "de Bruijn index" representation (instead of the nonexistent
struct).

(These corrections obviously shoudl not have been needed had I not got carried
away with throwing a bunch of structs in there... The pointers to "Term" are
obviously going to be pointing to new "Term" objects that will be used to read
in and assign as the untyped Plutus program is being parsed. I had not noticed
the error because (a) there are a lot of things to keep track of, and (b) the
code compiles fine with nonexistent structs acting like forward declarations
in C. Doh!)
IN 2021/11/28